Feature: Dual file loading
If no args are submitted to the argparse,
the program will scan the scraped
directory and give you a numbered
list of available json files.
You can then select and load from there

Or submit the path to the file you want to
load as you normally would with an argparser.